+++ client/src/design/index.ts
@@ -1,5 +1,6 @@
 export * from './themeTypes';
 export * from './defaultTheme';
 export * from './applyTheme';
+export * from './tokens';