+++ client/types/storybook.d.ts
@@ -1,0 +1,6 @@
+declare module '@storybook/react' {
+  export type Meta<TArgs = any> = any;
+  export type StoryObj<TArgs = any> = any;
+}
+
+